full client.rb testing
