+++ spec/support/mocks/callback.rb
@@ -1,0 +1,11 @@
+module Mocks
+  class CallbackTest
+    attr_reader :callback_value
+    def initialize
+      @callback_value = false
+    end
+    def callback_method(arg)
+      @callback_value = arg
+    end
+  end
+end
