+++ spec/support/mocks/response.rb
@@ -1,0 +1,9 @@
+module Mocks
+  class ResponseTest
+    attr_reader :response
+
+    def initialize(arg)
+      @response = arg
+    end
+  end
+end
